+++ src/components/input/Textarea.js
@@ -9,13 +9,19 @@ import theme from './theme.css';
 
 class Textarea extends PureComponent {
   render() {
-    const { className, error, helpText, inverse, ...others } = this.props;
+    const { className, error, helpText, inverse, warning, ...others } = this.props;
 
-    const classNames = cx(theme['wrapper'], className);
+    const classNames = cx(
+      theme['wrapper'],
+      {
+        [theme['has-error']]: error,
+        [theme['has-warning']]: warning,
+      },
+      className,
+    );
 
     const boxProps = pickBoxProps(others);
     const inputProps = {
-      error,
       inverse,
       ...omitBoxProps(others),
     };
@@ -23,7 +29,7 @@ class Textarea extends PureComponent {
     return (
       <Box className={classNames} {...boxProps}>
         <InputBase className={theme['textarea']} element="textarea" {...inputProps} />
-        <ValidationText error={error} help={helpText} inverse={inverse} />
+        <ValidationText error={error} help={helpText} inverse={inverse} warning={warning} />
       </Box>
     );
   }
@@ -38,6 +44,8 @@ Textarea.propTypes = {
   helpText: PropTypes.string,
   /** Boolean indicating whether the input should render as inverse. */
   inverse: PropTypes.bool,
+  /** The text to use as warning message below the input. */
+  warning: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
 };
 
 Textarea.defaultProps = {