+++ src/components/validationText/index.js
@@ -1,6 +1,7 @@
 import ValidationText from './ValidationText';
 import ErrorText from './ErrorText';
 import HelpText from './HelpText';
+import WarningText from './WarningText';
 
-export { ErrorText, HelpText };
+export { ErrorText, HelpText, WarningText };
 export default ValidationText;